refactor(projects): split ProjectDetail.tsx into four modules (#998)
ProjectDetail.tsx had grown to 828 lines (the issue recorded 805). Split
along the cohesion boundaries the file already had:

  projectDetailHelpers.ts       89  pure formatters + deriveChannels
  useProjectDetailActions.ts   215  mutating handlers + busy flags
  ProjectSourcesSection.tsx    106  sources table
  ProjectChannelsSection.tsx    79  channels palette
  ProjectDetail.tsx            466  composition + layout shell

Every file is now under the 500-line TS threshold.

Behaviour-preserving. The public API is unchanged — `ProjectDetailContent`
and its `ProjectDetail` alias still export from the same module, so all
eight importers and the seven ProjectDetail.*.test.tsx files are untouched.

The one real hazard here was hook ordering: `useProjectDetailActions` is
called unconditionally BEFORE the loading/error early returns, and its
internals run in the same relative position as the state and
`useGenerateArchivePlan` calls they replace, so React sees an identical hook
sequence. Its handlers guard on a possibly-undefined `project` accordingly.

ProjectChannelsSection returns null for an empty channel list, absorbing the
caller's `derivedChannels.length > 0` guard.

286 tests in src/features/projects pass, unchanged from the pre-split
baseline; typecheck and biome clean.
